@@ -8,9 +8,9 @@
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "axios": "^0.19.2",
-    "react": "^16.13.1",
+    "react": "^18.3.1",
     "react-beautiful-dnd": "^13.0.0",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.3.1",
     "react-flash-message": "^1.0.2",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
